Minhas modificações

+++ siso/src/main/java/com/gabrielpdev/siso/services/ItemMovimentoService.java
@@ -8,6 +8,7 @@ import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.sql.Timestamp;
 import java.util.List;
 import java.util.Optional;
 
@@ -17,37 +18,39 @@ public class ItemMovimentoService {
     @Autowired
     private ItemMovimentoRepository itemMovimentoRepository;
 
+    // Retorna todos os movimentos
     public List<ItemMovimento> findAll() {
         return itemMovimentoRepository.findAll();
     }
 
+    // Retorna um movimento por ID
     public ItemMovimento findById(Long id) {
-        Optional<ItemMovimento> itemMovimento = itemMovimentoRepository.findById(id);
-        return itemMovimentoRepository.findById(id).orElseThrow(() -> new ObjectNotFoundException("O itemMovimento {id: {"+id+"}} não foi encontrado"));
+        return itemMovimentoRepository.findById(id)
+                .orElseThrow(() -> new ObjectNotFoundException("O itemMovimento {id: " + id + "} não foi encontrado"));
     }
 
+    // Adiciona um novo movimento
     @Transactional
-    public void createItemMovimento(ItemMovimento itemMovimento) {
+    public ItemMovimento createItemMovimento(ItemMovimento itemMovimento) {
         itemMovimento.setId(null);
-        itemMovimentoRepository.save(itemMovimento);
+        return itemMovimentoRepository.save(itemMovimento); // Retorna o item salvo, que incluirá o ID gerado
     }
 
+    // Atualiza um movimento existente
     @Transactional
-    public void updateItemMovimento(ItemMovimento itemMovimento) {
-        ItemMovimento newItemMovimento = findById(itemMovimento.getId());
-        newItemMovimento.setCaixa(itemMovimento.getCaixa());
-        newItemMovimento.setDataHoraMovimento(itemMovimento.getDataHoraMovimento());
-        newItemMovimento.setDentista(itemMovimento.getDentista());
-        newItemMovimento.setFornecedor(itemMovimento.getFornecedor());
-        newItemMovimento.setDespesa(itemMovimento.getDespesa());
-        newItemMovimento.setReceita(itemMovimento.getReceita());
-        newItemMovimento.setValor(itemMovimento.getValor());
-        newItemMovimento.setOperacao(itemMovimento.getOperacao());
-        itemMovimentoRepository.save(newItemMovimento);
+    public ItemMovimento updateItemMovimento(ItemMovimento itemMovimento) {
+        if (!itemMovimentoRepository.existsById(itemMovimento.getId())) {
+            throw new ObjectNotFoundException("O itemMovimento {id: " + itemMovimento.getId() + "} não foi encontrado");
+        }
+
+        return itemMovimentoRepository.save(itemMovimento); // Atualiza e retorna o item atualizado
     }
 
-    public void deleteItemMovimentoById(Long id) {
-        findById(id);
+    // Deleta um movimento pelo ID
+    public void deleteItemMovimento(Long id) {
+        if (!itemMovimentoRepository.existsById(id)) {
+            throw new ObjectNotFoundException("O itemMovimento {id: " + id + "} não foi encontrado");
+        }
 
         try {
             itemMovimentoRepository.deleteById(id);
@@ -55,5 +58,19 @@ public class ItemMovimentoService {
             throw new DataBindingViolationException("O itemMovimento não pode ser deletado.");
         }
     }
-}
 
+    // Obtém todos os movimentos para um caixa específico
+    public List<ItemMovimento> getMovimentosPorCaixa(Long idCaixa) {
+        return itemMovimentoRepository.findByCaixaId(idCaixa);
+    }
+
+    // Retorna um movimento por ID
+    public ItemMovimento getItemMovimentoPorId(Long id) {
+        return itemMovimentoRepository.findById(id)
+                .orElseThrow(() -> new ObjectNotFoundException("O itemMovimento {id: " + id + "} não foi encontrado"));
+    }
+
+    public List<ItemMovimento> getMovimentosPorCaixasNoPeriodo(List<Long> caixaIds, Timestamp startDate, Timestamp endDate) {
+        return itemMovimentoRepository.findByCaixaIdInAndDataHoraMovimentoBetween(caixaIds, startDate, endDate);
+    }
+}
